Feature: Add local path based library generation
This change adds the ability to point to local files when generating
jsonnet libraries. This is done by attempting to match a scheme, and
assuming that crd paths without a scheme are local files.

The method used to get files (http.Get) would fail with files as it
requires a scheme to be set on requests.

This can be useful when generating libraries for locally vendored
deployment manifests that include CRDs.

---

This is intended to be upstreamed eventually as it allows more
flexibility in the approach to library generation.